Unify audio WAV encode/decode across base-nodes
The WAV encode/decode helpers were duplicated across three files:
  - `lib-audio-dsp.ts` defined Float32-based `encodeWav`/`decodeWav`.
  - `lib-audio-effects.ts` had a byte-for-byte copy with a comment
    explicitly saying "duplicated from lib-audio-dsp.ts".
  - `audio.ts` defined its own Int16-based `parseWavPcm16` /
    `buildWavFromSamples` plus private copies of `audioBytes`,
    `audioBytesAsync`, `audioRefFromBytes`, `concatBytes`, and `toBytes`.
  - `video.ts` re-implemented `audioBytes` and `toBytes` as well.

Extract the canonical helpers into a single module
`packages/base-nodes/src/lib/audio-wav.ts`:

  encodeWav(Float32Array, sampleRate, numChannels=1) -> Uint8Array
  decodeWav(audio) / tryDecodeWav(audio) -> WavData        (throws / null)
  parseWavBytes(bytes) -> WavData | null
  audioBytes(audio) / audioBytesAsync(audio, context)
  audioRefFromBytes(data, uri?) / audioRefFromWav(wav, uri?)
  concatBytes / toBytes / uriToPath

Every audio- or WAV-touching file now imports from this module; the
duplicate local definitions are removed. `NormalizeAudioNode` and
`AudioMixerNode` are refactored to operate in Float32 sample space via
the canonical `tryDecodeWav` + `encodeWav` pair (matching the rest of
`lib-audio-dsp.ts` and `lib-audio-effects.ts`) and drop the Int16-based
`parseWavPcm16` / `buildWavFromSamples` helpers entirely.

Net: 5 files changed, +38/-323 lines, one canonical source of truth.
