+++ packages/base-nodes/src/nodes/audio.ts
@@ -1,60 +1,24 @@
 import { BaseNode, prop } from "@nodetool/node-sdk";
-import type { AudioRef } from "@nodetool/node-sdk";
 import type { ProcessingContext } from "@nodetool/runtime";
 import { promises as fs } from "node:fs";
 import path from "node:path";
-
-type AudioRefLike = {
-  uri?: string;
-  data?: Uint8Array | string;
-};
+import {
+  audioBytes,
+  audioBytesAsync,
+  audioRefFromBytes,
+  audioRefFromWav,
+  concatBytes,
+  encodeWav,
+  toBytes,
+  tryDecodeWav,
+  uriToPath,
+  type WavData
+} from "../lib/audio-wav.js";
 
 type ImageLike = {
   data?: Uint8Array | string;
   uri?: string;
 };
-
-function toBytes(value: Uint8Array | string | undefined): Uint8Array {
-  if (!value) return new Uint8Array();
-  if (value instanceof Uint8Array) return value;
-  return Uint8Array.from(Buffer.from(value, "base64"));
-}
-
-function audioBytes(audio: unknown): Uint8Array {
-  if (!audio || typeof audio !== "object") return new Uint8Array();
-  const ref = audio as AudioRefLike;
-  if (ref.data) return toBytes(ref.data);
-  return new Uint8Array();
-}
-
-async function audioBytesAsync(audio: unknown, context?: ProcessingContext): Promise<Uint8Array> {
-  if (!audio || typeof audio !== "object") return new Uint8Array();
-  const ref = audio as AudioRefLike;
-  if (ref.data) return toBytes(ref.data);
-  if (typeof ref.uri === "string" && ref.uri) {
-    try {
-      if (context?.storage) {
-        const stored = await context.storage.retrieve(ref.uri);
-        if (stored !== null) return new Uint8Array(stored);
-      }
-      if (ref.uri.startsWith("file://")) {
-        return new Uint8Array(await fs.readFile(uriToPath(ref.uri)));
-      }
-      if (ref.uri.startsWith("http://") || ref.uri.startsWith("https://")) {
-        const response = await fetch(ref.uri);
-        return new Uint8Array(await response.arrayBuffer());
-      }
-    } catch {
-      return new Uint8Array();
-    }
-  }
-  return new Uint8Array();
-}
-
-function uriToPath(uriOrPath: string): string {
-  if (uriOrPath.startsWith("file://")) return uriOrPath.slice("file://".length);
-  return uriOrPath;
-}
 
 function dateName(name: string): string {
   const now = new Date();
@@ -66,25 +30,6 @@ function dateName(name: string): string {
     .replaceAll("%H", pad(now.getHours()))
     .replaceAll("%M", pad(now.getMinutes()))
     .replaceAll("%S", pad(now.getSeconds()));
-}
-
-function audioRefFromBytes(data: Uint8Array, uri?: string): AudioRef {
-  return {
-    type: "audio",
-    uri: uri ?? "",
-    data: Buffer.from(data).toString("base64")
-  };
-}
-
-function concatBytes(chunks: Uint8Array[]): Uint8Array {
-  const total = chunks.reduce((sum, c) => sum + c.length, 0);
-  const out = new Uint8Array(total);
-  let offset = 0;
-  for (const c of chunks) {
-    out.set(c, offset);
-    offset += c.length;
-  }
-  return out;
 }
 
 function getModelConfig(props: Record<string, unknown>): {
@@ -117,53 +62,6 @@ function hasProviderSupport(
     typeof context.streamProviderPrediction === "function" &&
     !!providerId &&
     !!modelId
-  );
-}
-
-function parseWavPcm16(
-  bytes: Uint8Array
-): { samples: Int16Array; headerSize: number } | null {
-  if (bytes.length < 44) return null;
-  const header = Buffer.from(bytes);
-  if (
-    header.toString("ascii", 0, 4) !== "RIFF" ||
-    header.toString("ascii", 8, 12) !== "WAVE"
-  ) {
-    return null;
-  }
-  const dataOffset = 44;
-  const pcm = bytes.slice(dataOffset);
-  if (pcm.length % 2 !== 0) return null;
-  return {
-    samples: new Int16Array(
-      pcm.buffer.slice(pcm.byteOffset, pcm.byteOffset + pcm.byteLength)
-    ),
-    headerSize: dataOffset
-  };
-}
-
-function buildWavFromSamples(
-  original: Uint8Array,
-  samples: Int16Array,
-  headerSize = 44
-): Uint8Array {
-  if (original.length >= headerSize) {
-    const out = new Uint8Array(headerSize + samples.byteLength);
-    out.set(original.slice(0, headerSize), 0);
-    out.set(
-      new Uint8Array(samples.buffer, samples.byteOffset, samples.byteLength),
-      headerSize
-    );
-    const view = new DataView(out.buffer);
-    view.setUint32(4, out.length - 8, true);
-    view.setUint32(40, samples.byteLength, true);
-    return out;
-  }
-  return new Uint8Array(
-    samples.buffer.slice(
-      samples.byteOffset,
-      samples.byteOffset + samples.byteLength
-    )
   );
 }
 
@@ -454,9 +352,8 @@ export class NormalizeAudioNode extends BaseNode {
   declare audio: any;
 
   async process(context?: ProcessingContext): Promise<Record<string, unknown>> {
-    const audio = this.audio;
-    const bytes = await audioBytesAsync(audio, context);
-    const wav = parseWavPcm16(bytes);
+    const bytes = await audioBytesAsync(this.audio, context);
+    const wav = tryDecodeWav({ data: bytes });
     if (!wav || wav.samples.length === 0) {
       return { output: audioRefFromBytes(bytes) };
     }
@@ -465,17 +362,14 @@ export class NormalizeAudioNode extends BaseNode {
     for (const sample of wav.samples) peak = Math.max(peak, Math.abs(sample));
     if (peak === 0) return { output: audioRefFromBytes(bytes) };
 
-    const gain = 32767 / peak;
-    const normalized = new Int16Array(wav.samples.length);
+    const gain = 1 / peak;
+    const normalized = new Float32Array(wav.samples.length);
     for (let i = 0; i < wav.samples.length; i += 1) {
-      normalized[i] = Math.max(
-        -32768,
-        Math.min(32767, Math.round(wav.samples[i] * gain))
-      );
+      normalized[i] = wav.samples[i] * gain;
     }
     return {
-      output: audioRefFromBytes(
-        buildWavFromSamples(bytes, normalized, wav.headerSize)
+      output: audioRefFromWav(
+        encodeWav(normalized, wav.sampleRate, wav.numChannels)
       )
     };
   }
@@ -1056,30 +950,29 @@ export class AudioMixerNode extends BaseNode {
     if (tracks.length === 0)
       return { output: audioRefFromBytes(new Uint8Array()) };
 
-    // If every track is a valid WAV PCM16 file, mix in sample space and
+    // If every track is a valid WAV file, mix in Float32 sample space and
     // emit a valid WAV so downstream nodes receive a playable file.
     const parsed = tracks.map((t) => ({
-      wav: parseWavPcm16(t.bytes),
+      wav: tryDecodeWav({ data: t.bytes }),
       bytes: t.bytes,
       volume: t.volume
     }));
     if (parsed.every((p) => p.wav !== null)) {
       const wavs = parsed as Array<{
-        wav: { samples: Int16Array; headerSize: number };
+        wav: WavData;
         bytes: Uint8Array;
         volume: number;
       }>;
       const len = Math.max(...wavs.map((p) => p.wav.samples.length));
-      const mixed = new Int16Array(len);
+      const mixed = new Float32Array(len);
       for (let i = 0; i < len; i += 1) {
         let total = 0;
         for (const p of wavs) total += (p.wav.samples[i] ?? 0) * p.volume;
-        const avg = total / wavs.length;
-        mixed[i] = Math.max(-32768, Math.min(32767, Math.round(avg)));
+        mixed[i] = total / wavs.length;
       }
       return {
-        output: audioRefFromBytes(
-          buildWavFromSamples(wavs[0].bytes, mixed, wavs[0].wav.headerSize)
+        output: audioRefFromWav(
+          encodeWav(mixed, wavs[0].wav.sampleRate, wavs[0].wav.numChannels)
         )
       };
     }
